Fetch Ansible steps logs and parse OK/Failed/Ignored task count
This commit extends the daily_matrix code to fetch the Ansible JSON
logs of the test steps, whenever they exist.

Example of JSON log file:
https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/periodic-ci-openshift-psap-ci-artifacts-release-4.7-upgrade-gpu-operator-e2e-upgrade-170/1407670009529896960/artifacts/gpu-operator-e2e-upgrade-170/2-postupgrade-gpu-operator-validate-deployment/artifacts/000__gpu-operator__wait_deployment/_ansible.log.json

From these Json logs, we parse the final array entry to extra the
number of tasks which passed (ok), failed or where skipped:
```
{
  "log_level": "info",
  "scope": "playbook",
  "stats": {
    "localhost": {
      "changed": 11,
      "failures": 0,
      "ignored": 0,
      "ok": 14,
      "rescued": 0,
      "skipped": 2,
      "unreachable": 0

  },
  "status": "finished"
}
```

These step counters are aggregated for each each tests, and display on
the HTML template page. For older tests, they are displayed as
tooltips.