Allow overriding test_matrix.ProwStep in test_spec.ProwStep
Most of our tests use `nightly` as test name, and so far this value
was configured once and for all, in the MatrixSpec.
This is the recommended way to proceed.

However, more complex tests require multiple steps, like the GPU
Operator upgrade testing where we have 3 steps.

This commits allows overriding the Prow Step name for each of the
tests. If the field is left empty, the Matrix Prow Step name is used.